# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ParchiFlow - Universal SME Ledger Intelligence & Collection Automation OS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ledger:
    get:
      operationId: getLedgerEntries
      tags: [ledger]
      summary: Get ledger entries
      parameters:
        - name: party_id
          in: query
          schema:
            type: integer
        - name: entry_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
    post:
      operationId: createLedgerEntry
      tags: [ledger]
      summary: Create a ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerEntryBody"
      responses:
        "201":
          description: Created ledger entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
  /ledger/{id}:
    get:
      operationId: getLedgerEntry
      tags: [ledger]
      summary: Get a ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"

  # MONEY EVENTS
  /money-events:
    get:
      operationId: getMoneyEvents
      tags: [money-events]
      summary: Get money events
      parameters:
        - name: party_id
          in: query
          schema:
            type: integer
        - name: event_type
          in: query
          schema:
            type: string
        - name: reconciliation_status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Money events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoneyEvent"
    post:
      operationId: createMoneyEvent
      tags: [money-events]
      summary: Create a money event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoneyEventBody"
      responses:
        "201":
          description: Created money event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyEvent"
  /money-events/{id}:
    get:
      operationId: getMoneyEvent
      tags: [money-events]
      summary: Get a money event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Money event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyEvent"

  # OUTSTANDINGS
...
  /follow-ups:
    get:
      operationId: getFollowUps
      tags: [follow-ups]
      summary: Get follow-ups
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: party_id
          in: query
          schema:
            type: integer
        - name: tab
          in: query
          schema:
            type: string
            description: "due_today | overdue | promised_today | high_priority | disputed | all"
      responses:
        "200":
          description: Follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowUp
      tags: [follow-ups]
      summary: Create a follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpBody"
      responses:
        "201":
          description: Created follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /follow-ups/{id}:
    put:
      operationId: updateFollowUp
      tags: [follow-ups]
      summary: Update a follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFollowUpBody"
      responses:
        "200":
          description: Updated follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"

  # REMINDERS
...
components:
  schemas:
...
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        partyId:
          type: integer
          nullable: true
        partyName:
          type: string
          nullable: true
        entryType:
          type: string
        voucherNumber:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
        amount:
          type: number
        debitAmount:
          type: number
          nullable: true
        creditAmount:
          type: number
          nullable: true
        entryDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
          nullable: true
        narration:
          type: string
          nullable: true
        status:
          type: string
        reconciliationStatus:
          type: string
        sourceType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, entryType, amount, entryDate, status, reconciliationStatus, createdAt]
    CreateLedgerEntryBody:
      type: object
      properties:
        partyId:
          type: integer
          nullable: true
        entryType:
          type: string
        amount:
          type: number
        entryDate:
          type: string
        dueDate:
          type: string
          nullable: true
        narration:
          type: string
          nullable: true
        voucherNumber:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
      required: [entryType, amount, entryDate]

    # MONEY EVENT SCHEMAS
    MoneyEvent:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        sourceType:
          type: string
        eventType:
          type: string
        amount:
          type: number
        direction:
          type: string
        eventDate:
          type: string
          format: date-time
        partyId:
          type: integer
          nullable: true
        partyName:
          type: string
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        utr:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
        narration:
          type: string
          nullable: true
        confidenceScore:
          type: number
        reconciliationStatus:
          type: string
        rawInput:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, sourceType, eventType, amount, direction, eventDate, confidenceScore, reconciliationStatus, createdAt]
    CreateMoneyEventBody:
      type: object
      properties:
        sourceType:
          type: string
        eventType:
          type: string
        amount:
          type: number
        direction:
          type: string
        eventDate:
          type: string
        partyId:
          type: integer
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        utr:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
        narration:
          type: string
          nullable: true
        rawInput:
          type: string
          nullable: true
      required: [sourceType, eventType, amount, direction, eventDate]

    # OUTSTANDING SCHEMAS
...
    FollowUp:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        partyId:
          type: integer
        partyName:
          type: string
          nullable: true
        partyMobile:
          type: string
          nullable: true
        outstandingId:
          type: integer
          nullable: true
        outstandingAmount:
          type: number
          nullable: true
        agingDays:
          type: integer
          nullable: true
        followUpType:
          type: string
        status:
          type: string
        note:
          type: string
          nullable: true
        promisedPaymentDate:
          type: string
          format: date-time
          nullable: true
        promisedAmount:
          type: number
          nullable: true
        nextFollowUpAt:
          type: string
          format: date-time
          nullable: true
        lastReminderAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, partyId, followUpType, status, createdAt]
    CreateFollowUpBody:
      type: object
      properties:
        partyId:
          type: integer
        outstandingId:
          type: integer
          nullable: true
        followUpType:
          type: string
        note:
          type: string
          nullable: true
        promisedPaymentDate:
          type: string
          nullable: true
        promisedAmount:
          type: number
          nullable: true
        nextFollowUpAt:
          type: string
          nullable: true
      required: [partyId, followUpType]
    UpdateFollowUpBody:
      type: object
      properties:
        status:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        promisedPaymentDate:
          type: string
          nullable: true
        promisedAmount:
          type: number
          nullable: true
        nextFollowUpAt:
          type: string
          nullable: true
      required: []

    # REMINDER SCHEMAS
...